% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.didImputation}
\alias{coef.didImputation}
\title{coef method for object of class "didImputation"}
\usage{
\method{coef}{didImputation}(object, ...)
}
\arguments{
\item{object}{an object for which the extraction of model coefficients is
    meaningful.}

\item{...}{other arguments.}
}
\description{
coef method for object of class "didImputation"
}
