% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{computeWeightControl}
\alias{computeWeightControl}
\title{One iteration of weight}
\usage{
computeWeightControl(dt, w_i, denom, c)
}
\arguments{
\item{dt}{data.table.}

\item{w_i}{Character. Weight to demean.}

\item{denom}{Numeric. Relative size of untreated population.}

\item{c}{Character. Name of the control to demean on.}
}
\value{
Void. Mutate the weight vector in the original data.table.
}
\description{
Single iteration of weight demean w.r.t. continuous control.
}
