% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{computeWeights}
\alias{computeWeights}
\title{Compute weights}
\usage{
computeWeights(s, ...)
}
\arguments{
\item{s}{a "didImputation" model object}

\item{...}{other arguments.}
}
\value{
a didImputation extended model
}
\description{
Compute weights
}
