% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardErrors.R
\name{computeWeightSE}
\alias{computeWeightSE}
\title{Weight specific standard error}
\usage{
computeWeightSE(w_i, dt, s, ...)
}
\arguments{
\item{w_i}{Character. Name of the weight.}

\item{dt}{data.table. Estimation table.}

\item{s}{Object of class "didImputation".}

\item{...}{other arguments.}
}
\description{
Weight specific standard error
}
