% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{did_simulated}
\alias{did_simulated}
\title{Simulated DiD}
\format{
A data frame with 1500 rows and 4 variables:
\describe{
\item{i}{Unid id}
\item{t}{Calendar time}
\item{g}{Date of treatment, Inf for never treated}
\item{y}{Dependent variable}
...
}
}
\usage{
did_simulated
}
\description{
A dataset of simulated panel data with multiple treatment groups.
}
\keyword{datasets}
