% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/didImputation.R
\name{didImputation}
\alias{didImputation}
\title{DiD regression with imputation method}
\usage{
didImputation(
  y0,
  data,
  cohort,
  nevertreated.value = Inf,
  unit = NULL,
  time = NULL,
  td = FALSE,
  w = NULL,
  coef = -Inf:Inf,
  OATT = FALSE,
  with.se = TRUE,
  tol = 1e-06,
  maxit = 100,
  mutatedata = FALSE,
  verbose = 0,
  effective.sample = 30
)
}
\arguments{
\item{y0}{Formula. model for Y(0). This is the full model without leads and lags used to predict the counterfactual outcome.}

\item{data}{A data frame or data.table}

\item{cohort}{Character. Time of treatment identifier. By default it expect \code{Inf} for never treated units. You can override this with \code{nevertreated.value}}

\item{nevertreated.value}{Any, default is \code{Inf}. Value encoding the cohort of never treated units.}

\item{unit}{Character, default is \code{NULL}. Unit identifier. If \code{NULL}, it default to the first fixed effect.}

\item{time}{Character, default is \code{NULL}. Time period identifier. If \code{NULL}, it default to the second fixed effect.}

\item{td}{Logical or Character, default is \code{FALSE}. Triple difference estimation. If \code{TRUE}, the third fixed effect in \code{y0} will be interpreted as the additional group. If Character, then the corresponding column will be interpreted as the new group.}

\item{w}{Numerical vector or Variable name. Default is \code{NULL}. Sample weights.}

\item{coef}{R Expression of leads and lags (eg \code{coef = -5:8}). By default estimates whole dynamic effect. See Details.}

\item{OATT}{Logical, default is \code{FALSE}. If \code{TRUE} then the overall average treatment effect is computed instead of ATT at each horizon.}

\item{with.se}{Logical, default is \code{TRUE}. Should standard errors be reported}

\item{tol}{Numeric, default is \code{1e-6}. Tolerance level for weights convergence.}

\item{maxit}{Numeric, default is \code{100}. Maximum number of iterations for computing weights}

\item{mutatedata}{Logical, default is \code{FALSE}. USE AT YOUR OWN RISK. This option will modify your data in place instead
of making a copy of it which could be useful with large dataset if your RAM is limited.
It may result in lost observations and bugs.}

\item{verbose}{Numeric, default is \code{0}. Level of verbosity.}

\item{effective.sample}{Numeric, default is \code{30}. Effective sample size under which the function will throw a warning. See details.}
}
\value{
A didImputation object with the results of the imputation estimation.
\item{data}{Data used for estimation.}
\item{convergence_time}{User and CPU time for weights convergence.}
\item{pvalue}{p-value for the positive horizon estimates.}
\item{coeftable}{Table of regression coefficients.}
\item{wald}{Wald statistic of the pre-trend regression.}
\item{coefs}{Average treatment effects from the imputation procedure.}
\item{niterations}{Number of iterations it took to compute the weights.}
\item{pre_trends}{\code{fixest} regression object for the pre-trends estimation.}
}
\description{
DiD regression with imputation method
}
\details{
See below for additional details on some arguments.
}
\section{Estimated coefficients}{


By default, the function will estimate all coefficients available. You can
customize this behavior with the \code{coef} option. The option takes an R
expression of the form \code{-{leads}:{lags}}. The default behavior is to estimate
all available coefficients such that \code{coef = -Inf:Inf}.

For pre-trend coefficients, by default the function will set the greatest
leads as the reference group if \code{-Inf} is set.
}

\section{Effective sample a.k.a. Herfindahl condition}{


Borusyak, K., Jaravel, X., & Spiess, J. give a condition on weights such that
consistency of estimators holds. This condition states that the Herfindahl
index of weights must converge to zero. Another interpretation is that the
inverse of the index is a measure of 'effective sample size'. Authors
recommand an effective sample size of at least 30. If the effective sample
size is lower, a warning will be thrown.
}

\section{Cohort}{


The \code{cohort} argument is the date of treatment of the observation. By default
it expect \code{Inf} for never treated individuals. You can override this behavior
by setting another value to \code{nevertreated.value}.
}

\section{Standard-errors}{


The standard errors are computed using an alternating projection method. You
can tweek its meta-parameter by changing \code{tol} and \code{maxit}.
}

\examples{
#Load example data
data(did_simulated)

# Estimate the overall average treatment effect on treated and all available pre-trends
didImputation(y ~ 0 | i + t,
              cohort = 'g',
              OATT = TRUE,
              data = did_simulated)

# Estimate the full dynamic model
didImputation(y ~ 0 | i + t,
              cohort = 'g',
              data = did_simulated)

# Estimate positive (lags) coefficients
didImputation(y ~ 0 | i + t,
              cohort = 'g',
              coef = 0:Inf,
              data = did_simulated)

# Estimate first 3 post treatment coefficients
didImputation(y ~ 0 | i + t,
              cohort = 'g',
              coef = 0:2,
              data = did_simulated)

# Return only point estimates
didImputation(y ~ 0 | i + t,
              cohort = 'g',
              with.se = FALSE,
              data = did_simulated)


}
\references{
Borusyak, K., Jaravel, X., & Spiess, J. (2021). Revisiting event study designs: Robust and efficient estimation. Working paper.

Stata implementation from the authors:

did_imputation (\url{https://github.com/borusyak/did_imputation})

Another R implementation using sparse matrix inversion:
{didimputation} (\url{https://github.com/kylebutts/didimputation})
}
\author{
Antoine Mayerowitz

Maxime Gravoueille
}
