% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{didplot}
\alias{didplot}
\title{DiD imputation plot}
\usage{
didplot(object, type = "default", ci = 0.95, ...)
}
\arguments{
\item{object}{an object of class DidImputation.}

\item{type}{Character. Type of plot. Can be either 'default' or 'IR' for
impulse response style.}

\item{ci}{Numeric. Confidence interval( 0.95 by default).}

\item{...}{additional arguments.}
}
\value{
A ggplot2 object
}
\description{
Takes a didImputation estimation and return a graph of estimated
parameters(for dynamic effects only).
}
\examples{
# standard plot
res <- didImputation(y ~ 0 | i + t,
              cohort = 'g',
              data = did_simulated)

didplot(res)

# Plot with 99\% confidence interval

didplot(res, ci = 0.99)

# Plot with Impulse response function style

didplot(res, type = 'IR')


}
