% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generateDidData}
\alias{generateDidData}
\title{Generate longitudinal data with staggered treatment}
\usage{
generateDidData(
  i,
  t,
  hdfe = TRUE,
  control = TRUE,
  attrition.rate = 0,
  treatment = ~d * (k + 1)
)
}
\arguments{
\item{i}{Numeric. Number of units}

\item{t}{Numeric. Number of periods}

\item{hdfe}{Logical. Should high dimensional fixed effect be added?}

\item{control}{Logical. Should continuous control be added?}

\item{attrition.rate}{Numerical. Default is 0. Probability of missing observation.}

\item{treatment}{one sided formula, default is  ~ d * (k + 1). See details.}
}
\value{
a data.table with columns
\item{i}{Unit identifier.}
\item{t}{Calendar time.}
\item{g}{Time of treatment.}
\item{a}{Individual fixed effect. ~N(0,1)}
\item{b}{Time fixed effect. ~N(0,1)}
\item{x1}{A continuous, normally distributed control. ~N(0,2)}
\item{hdfe}{A random effect with \{i\}/50 modalities. ~ N(0,1)}
\item{e}{Residual. ~ N(0,2)}
\item{hdfefactor}{A fixed effect of size \{i\}/50.}
\item{k}{Relative time to treatment.}
\item{d}{Dummy variable defined as k >= 0}
\item{h}{Randomly generated variable at the level of the cohort.}
\item{y}{Outcome.}
\item{true_effect}{True average treatment effect on the treated.}
}
\description{
Generate longitudinal data with staggered treatment
}
\section{Treatment effect formula}{


By default the treatment effect is \code{~ d * (k + 1)} where \code{d} is an internal
dummy equal to one if observation is treated and \code{k} is the relative time to
event. User can define its own formula to create complex treatment and test
the model effectiveness.

User has access to the following hidden variables:
\itemize{
\item \code{h}: a cohort-wise randomly generated number
\item \code{g}: Cohort identifier
\item \code{a}: Individual fixed effect
\item \code{b}: Time fixed effect
\item \code{t}: Calendar time
}
}

\section{Default outcome}{


The default data generating process is:

\code{y ~ a + b + hdfe + 2*x1 + d * (k + 1) + e}

Variable description is available in previous section.
}

\examples{
dt <- generateDidData(100,10)
head(dt)

}
