% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{iterateWeight}
\alias{iterateWeight}
\title{Recursively estimate weights until convergence}
\usage{
iterateWeight(params, dt, i, w)
}
\arguments{
\item{params}{List. Object of class DiDImputation.}

\item{dt}{data.table. Internal regression table.}

\item{i}{Integer. Recursion index.}

\item{w}{Character. Name of the weight to estimate.}
}
\value{
Integer. The recursion index (i.e, the number of function calls)
}
\description{
Recursively estimate weights until convergence
}
