% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{parseCoef}
\alias{parseCoef}
\title{Parse coef argument expression}
\usage{
parseCoef(coef, OATT, k)
}
\arguments{
\item{coef}{Expression. The amount of leads and lags allowing for Inf.}

\item{OATT}{Logical. Should the}

\item{k}{Numeric vector. Relative time to event}
}
\value{
Numeric vector of length 2.
}
\description{
Parse coef argument expression
}
