% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parseControls}
\alias{parseControls}
\title{Parse control variables}
\usage{
parseControls(f, ...)
}
\arguments{
\item{f}{Formula.}

\item{...}{additional arguments.}
}
\value{
a vector of expressions.
}
\description{
Parse control variables
}
