% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parseFEs}
\alias{parseFEs}
\title{Parse Fixed Effects}
\usage{
parseFEs(f, ...)
}
\arguments{
\item{f}{Formula.}

\item{...}{additional arguments.}
}
\value{
a character vector with fixed effects.
}
\description{
Extract fixed effects from formula
}
