% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.didImputation}
\alias{plot.didImputation}
\title{DiD imputation plot}
\usage{
\method{plot}{didImputation}(...)
}
\arguments{
\item{...}{additional arguments.}
}
\value{
A ggplot2 object
}
\description{
Takes a didImputation estimation and return a graph of estimated
parameters(for dynamic effects only).
}
\examples{
# standard plot
res <- didImputation(y ~ 0 | i + t,
              cohort = 'g',
              data = did_simulated)

didplot(res)

# Plot with 99\% confidence interval

didplot(res, ci = 0.99)

# Plot with Impulse response function style

didplot(res, type = 'IR')


}
