% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare}
\alias{prepare}
\title{Create variables for treatment status and horizon.}
\usage{
prepare(s)
}
\arguments{
\item{s}{A list with the arguments to didImputation.}
}
\value{
Augmented list with configured arguments.
}
\description{
Create variables for treatment status and horizon.
}
