% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preTrend.R
\name{preTrend}
\alias{preTrend}
\title{Test of parallel trends}
\usage{
preTrend(s, ...)
}
\arguments{
\item{s}{an object of class "didImputation".}

\item{...}{additional arguments.}
}
\value{
A fixest regression result with pre-trend coefficients.
}
\description{
Test of parallel trends
}
