% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.didImputation}
\alias{print.didImputation}
\title{Print didImputation result}
\usage{
\method{print}{didImputation}(x, ...)
}
\arguments{
\item{x}{an object of class "didImputation".}

\item{...}{additional arguments.}
}
\description{
Print didImputation result
}
