% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.didImputation}
\alias{print.summary.didImputation}
\title{Print method for summary of class "didImputation"}
\usage{
\method{print}{summary.didImputation}(x, digits = 3, ...)
}
\arguments{
\item{x}{A list of class "summary.didImputation"}

\item{digits}{Decimal precision in output.}

\item{...}{additional arguments.}
}
\value{
An estimation table
}
\description{
Print method for summary of class "didImputation"
}
