% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{runImputation}
\alias{runImputation}
\title{Imputation estimation.}
\usage{
runImputation(s, ...)
}
\arguments{
\item{s}{an object of class "didImputation"}

\item{...}{additional arguments.}
}
\value{
an object of class "didImputation" mutated with imputed values.
}
\description{
Imputation estimation.
}
