% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subsetData}
\alias{subsetData}
\title{Subset data according to function arguments}
\usage{
subsetData(data, s)
}
\arguments{
\item{data}{A data frame.}

\item{s}{A list containing model parameters.}
}
\value{
A data frame.
}
\description{
Subset data according to function arguments
}
