% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.didImputation}
\alias{summary.didImputation}
\title{DiD imputation summary result}
\usage{
\method{summary}{didImputation}(object, ...)
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
DiD imputation summary result
}
